// src/app/api/model/vendor.ts
/**
 * SmartLedger API
 *
 * Contact: support@smartledger.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Invoice } from './invoice';
import { Organization } from './organization';


export interface Vendor { 
    id?: string;
    orgId?: string | null;
    name?: string | null;
    email?: string | null;
    phone?: string | null;
    category?: string | null;
    createdAt?: string | null;
    invoices?: Array<Invoice> | null;
    org?: Organization;
}