// src/app/api/model/organization.ts
/**
 * SmartLedger API
 *
 * Contact: support@smartledger.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Invoice } from './invoice';
import { User } from './user';
import { CashflowPrediction } from './cashflowPrediction';
import { Vendor } from './vendor';


export interface Organization { 
    id?: string;
    name?: string | null;
    industry?: string | null;
    gstNumber?: string | null;
    country?: string | null;
    plan?: string | null;
    createdAt?: string | null;
    cashflowPredictions?: Array<CashflowPrediction> | null;
    invoices?: Array<Invoice> | null;
    users?: Array<User> | null;
    vendors?: Array<Vendor> | null;
}