// src/app/api/model/vendorRequest.ts
/**
 * SmartLedger API
 *
 * Contact: support@smartledger.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VendorRequest { 
    name?: string | null;
    email?: string | null;
    phone?: string | null;
    category?: string | null;
}